Fixed boot error with Application file and Venue missing @ID
