+++ BlackJackets API/src/main/java/BlackJackets/BlackJackets/controllers/GigController.java
@@ -1,7 +1,7 @@
-package controllers;
+package BlackJackets.BlackJackets.controllers;
 
-import data.GigRepository;
-import models.Gig;
+import BlackJackets.BlackJackets.data.GigRepository;
+import BlackJackets.BlackJackets.models.Gig;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,7 +16,7 @@ public class GigController {
     private GigRepository gigRepository;
 
     @PostMapping("Gig")
-    public Gig processCreateGig(@RequestBody String gigName){
+    public Gig processCreateGig(@RequestBody String gigName) {
         return new Gig(gigName);
     }
