+++ BlackJackets API/src/main/java/BlackJackets/BlackJackets/data/GigRepository.java
@@ -1,8 +1,10 @@
-package data;
+package BlackJackets.BlackJackets.data;
 
-import models.Gig;
+
+import BlackJackets.BlackJackets.models.Gig;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+
 
 @Repository
 public interface GigRepository extends CrudRepository<Gig, Long> {
