+++ BlackJackets API/src/main/java/BlackJackets/BlackJackets/models/Gig.java
@@ -1,7 +1,30 @@
 package BlackJackets.BlackJackets.models;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
 
 @Entity
 public class Gig {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    private String name;
+
+    public Gig(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
 }
