+++ BlackJackets API/src/main/java/BlackJackets/BlackJackets/models/Venue.java
@@ -1,7 +1,12 @@
 package BlackJackets.BlackJackets.models;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
 
 @Entity
 public class Venue {
+    @Id
+    @GeneratedValue
+    private Long id;
 }
