--- BlackJackets API/src/main/java/models/Gig.java
@@ -1,30 +1,0 @@
-package models;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-
-@Entity
-public class Gig {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    private String name;
-
-    public Gig(String name) {
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
